Changed the LVGSolverPopulationRequest. Removed column density, and added MoleculeAbundanceRatios and ConstantNmolBydVdR (should be ConstantNpartnerBydVdR). Fixed a bug in the OpticalDepthProvider.

+++ Scripts/compare2Results.m
@@ -4,41 +4,63 @@ function compare2Results ()
 Result1 = WorkspaceHelper.GetLVGResultFromWorkspace('Radex - HCO+ - T60K');
 Result2 = WorkspaceHelper.GetLVGResultFromWorkspace('HCO+ - T60K');
 
+dvdrKmParsecIndices = 1:numel(Result.OriginalRequest.VelocityDerivative);
+%temperatureIndices = 1:numel(Result.OriginalRequest.Temperature);
+temperatureIndices = find(Result.OriginalRequest.Temperature==600);
+%densityIndices = 1:numel(Result.OriginalRequest.CollisionPartnerDensities);
+collisionPartnerDensityIndices = find(Result.OriginalRequest.CollisionPartnerDensities==3e5);
+moleculeAbundanceIndices = 1:numel(Result.OriginalRequest.MoleculeAbundanceRatios);
+
+
 CompareType = ComparisonTypeCodes.Tau;
+
+SaveImages = 1;
+GenerateReport = 1;
 
 %%%%% DON'T CHANGE CODE BELOW HERE
 
-p = mfilename('fullpath'); %returns path of current script.
-ScriptsDirectory = fileparts(p);
-picsPath = fullfile(ScriptsDirectory, '..', 'Results', 'RadexCompareOutput', 'pics');
-[status, message, messageid] = rmdir(picsPath, 's');
-mkdir(picsPath);
+figure;
+
+if SaveImages
+    picsPath = FileIOHelper.ComparisonPicsOutputPath();
+    [status, message, messageid] = rmdir(picsPath, 's');
+    mkdir(picsPath);
+end
 
 DiffGrid = zeros(numel(Result1.OriginalRequest.Temperature),numel(Result1.OriginalRequest.CollisionPartnerDensities));
 
-for dvdrKmParsec = 1:numel(Result1.OriginalRequest.VelocityDerivative)
+for dvdrKmParsec = dvdrKmParsecIndices
+    for temp = temperatureIndices
+        for molAbundance=moleculeAbundanceIndices
+            for dens = collisionPartnerDensityIndices
+                
+                %Compare results using a graph
+                picFileName = sprintf('Temperature%g_%g.jpg',Result1.OriginalRequest.Temperature(temp), dens);
+                FileName = fullfile(picsPath, picFileName);
+                %
+                [cmp1, cmp1Title, cmp2, cmp2Title, YLabel, FixYAxis] = runParameters (Result1, Result2, CompareType, temp, dens, dvdrKmParsec,molAbundance);
+                
+                Data = cell(1,2);
+                plotArguments = cell(1,2);
+                Data{1} = cmp1;
+                Data{2} = cmp2;
+                plotArguments{1} = {'DisplayName', cmp1Title};
+                plotArguments{2} = {'DisplayName', cmp2Title};
 
-    for temp = 1:numel(Result1.OriginalRequest.Temperature)
-        
-        for dens = 1:numel(Result1.OriginalRequest.CollisionPartnerDensities)
-            
-            %Compare results using a graph
-            picFileName = sprintf('Temperature%g_%g.jpg',Result1.OriginalRequest.Temperature(temp), dens);
-            FileName = fullfile(picsPath, picFileName);
-            %
-      
-            [cmp1, cmp1Title, cmp2, cmp2Title, YLabel, FixYAxis] = runParameters (Result1, Result2, CompareType, temp, dens, dvdrKmParsec);
-      
-            Scripts.Compare2Results(cmp1, cmp1Title, cmp2, cmp2Title, YLabel, FixYAxis, Result1.OriginalRequest.CollisionPartnerDensities(dens), Result1.OriginalRequest.Temperature(temp), ...
-                Result1.OriginalRequest.VelocityDerivative(dvdrKmParsec), FileName);
-            
-            diff = Scripts.PopulationDiff(cmp1, cmp2);
-            
-            DiffGrid(temp,dens) = 100*max(diff);
-            if any (isnan(diff)) %|| RadexConverged == 0
-                'Error'
+                %(Data, PlotArguments, YLabel, XLabel,YRange, YAxisLog, Title, FileName)
+                
+                Scripts.CompareResults(Data, plotArguments, YLabel, );
+                
+                Scripts.CompareResults(cmp1, cmp1Title, cmp2, cmp2Title, YLabel, FixYAxis, Result1.OriginalRequest.CollisionPartnerDensities(dens), Result1.OriginalRequest.Temperature(temp), ...
+                    Result1.OriginalRequest.VelocityDerivative(dvdrKmParsec), FileName);
+                
+                diff = Scripts.PopulationDiff(cmp1, cmp2);
+                
+                DiffGrid(temp,dens) = 100*max(diff);
+                if any (isnan(diff)) %|| RadexConverged == 0
+                    'Error'
+                end
             end
-            
         end
     end
 end
@@ -48,7 +70,7 @@ generateComparisonResport(Result1.OriginalRequest.Temperature, Result1.OriginalR
 end
 
 
-function [cmp1, cmp1Title, cmp2, cmp2Title, YLabel, FixYAxis] = runParameters (Result1, Result2, ComparisonType, tempIndex, densIndex, dvdrKmParsecIndex)
+function [cmp1, cmp1Title, cmp2, cmp2Title, YLabel, FixYAxis] = runParameters (Result1, Result2, ComparisonType, tempIndex, densIndex, dvdrKmParsecIndex,molAbundanceIndex)
 
 cmp1Title = RunTypeCodes.ToStringGUIFormat(Result1.OriginalRequest.RunTypeCode);
 cmp2Title = RunTypeCodes.ToStringGUIFormat(Result2.OriginalRequest.RunTypeCode);
@@ -56,15 +78,15 @@ cmp2Title = RunTypeCodes.ToStringGUIFormat(Result2.OriginalRequest.RunTypeCode);
 switch ComparisonType
     case ComparisonTypeCodes.Population
         
-        cmp1 = Result1.Population(:,tempIndex,densIndex,dvdrKmParsecIndex);
-        cmp2 = Result2.Population(:,tempIndex,densIndex,dvdrKmParsecIndex);
+        cmp1 = Result1.Population(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
+        cmp2 = Result2.Population(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
         YLabel = 'x - Fractional population';
         FixYAxis = true;
         
     case ComparisonTypeCodes.Intensities
         
-        cmp1 = Result1.Intensities(:,tempIndex,densIndex,dvdrKmParsecIndex);
-        cmp2 = Result2.Intensities(:,tempIndex,densIndex,dvdrKmParsecIndex);
+        cmp1 = Result1.Intensities(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
+        cmp2 = Result2.Intensities(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
         YLabel = 'Intensities';
         FixYAxis = false;
 
@@ -75,15 +97,15 @@ switch ComparisonType
         
     case ComparisonTypeCodes.Beta
                 
-        cmp1 = Result1.FinalBetaCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex);
-        cmp2 = Result2.FinalBetaCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex);
+        cmp1 = Result1.FinalBetaCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
+        cmp2 = Result2.FinalBetaCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
         YLabel = 'Beta';
         FixYAxis = false;        
                 
     case ComparisonTypeCodes.Tau
                 
-        cmp1 = Result1.FinalTauCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex);
-        cmp2 = Result2.FinalTauCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex);
+        cmp1 = Result1.FinalTauCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
+        cmp2 = Result2.FinalTauCoefficients(:,tempIndex,densIndex,dvdrKmParsecIndex,molAbundanceIndex);
         YLabel = 'Tau';
         FixYAxis = false;
